feat(checkout): INT-3112 Merge Bolt strategies
