+++ src/payment/strategies/bolt/bolt-payment-initialize-options.ts
@@ -1,0 +1,7 @@
+export default interface BoltPaymentInitializeOptions {
+  /**
+   * When true, BigCommerce's checkout will be used
+   * otherwise Bolt's full checkout take over will be assumed
+   */
+  useBigCommerceCheckout?: boolean;
+}
